% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accept_k_means.R
\name{accept_k_means}
\alias{accept_k_means}
\title{After using emptydrops, set FDR_multi<-1 for cells above the ambiguous area, and set FDR_multi<-0 in cells below the lower line.}
\usage{
accept_k_means(e_multi.out, lower_atac, lower_rna)
}
\arguments{
\item{e_multi.out}{Dataframe with columns for Total_RNA, Total_chromatin, rownames, barhop_amb_tent_chromatin, barhop_amb_tent_rna, FDR_multi}

\item{lower_atac}{A single number to act as the value of the lower parameter in ATAC}

\item{lower_rna}{A single number to act as the value of the lower parameter in RNA}
}
\value{
Dataframe with columns for Total_RNA, Total_chromatin, rownames, barhop_amb_tent_chromatin,
barhop_amb_tent_rna, FDR_multi, FDR, above_ambiguous, k_means
}
\description{
After using emptydrops, set FDR_multi<-1 for cells above the ambiguous area, and set FDR_multi<-0 in cells below the lower line.
}
