% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggr_p_values.R
\name{aggr_p_values}
\alias{aggr_p_values}
\title{Aggregates the p-values from the two modalities (RNA and epigenomics) and then corrects them using the Benjamini-Hochberg method}
\usage{
aggr_p_values(e.out1, e.out2)
}
\arguments{
\item{e.out1}{a DataFrame (in the format of the emptyDrops outputs) with the following components:
Total:
Integer, the total count for each barcode.
LogProb:
Numeric, the log-probability of observing the barcode's count vector under the null model.
PValue:
Numeric, the Monte Carlo p-value against the null model.
Limited:
Logical, indicating whether a lower p-value could be obtained by increasing niters.
FDR:
Numeric, the p-values corrected using the Benjamini-Hochberg method}

\item{e.out2}{a DataFrame (in the format of the emptyDrops outputs) like e.out1}
}
\value{
a DataFrame (in the format of the emptyDrops outputs) with the following components:
Total_RNA:
Integer, the total count for each barcode.
LogProb_RNA:
Numeric, the log-probability of observing the barcode's count vector under the null model.
PValue_RNA:
Numeric, the Monte Carlo p-value against the null model.
Limited_RNA:
Logical, indicating whether a lower p-value could be obtained by increasing niters.
FDR_RNA:
Numeric, the p-values corrected using the Benjamini-Hochberg method
Total_chromatin:
Integer, the total count for each barcode.
LogProb_chromatin:
Numeric, the log-probability of observing the barcode's count vector under the null model.
PValue_chromatin:
Numeric, the Monte Carlo p-value against the null model.
Limited_chromatin:
Logical, indicating whether a lower p-value could be obtained by increasing niters.
FDR_chromatin:
Numeric, the p-values corrected using the Benjamini-Hochberg method
PValue_multi:
Numeric, the p-values of the droplets after aggregating PValue_chromatin and PValue_RNA for each droplet
FDR_multi:
Numeric, the result of the correction of PValue_multi using the Benjamini-Hochberg method
}
\description{
Aggregates the p-values from the two modalities (RNA and epigenomics) and then corrects them using the Benjamini-Hochberg method
}
