% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher.R
\name{balanced_aggr}
\alias{balanced_aggr}
\title{Aggregate p-values using the average of the Fisher aggregation value and the mean aggregation method.}
\usage{
balanced_aggr(pvalues, r)
}
\arguments{
\item{pvalues}{Vector of real numbers in the range from 0 to 1.}

\item{r}{r Real number for the kind of mean which will be used to aggregate the p-values.
For instance, r=1 is the arithmetic mean, r=-1 is the harmonic mean etc.}
}
\value{
Real number of the aggregated p-value.
}
\description{
Aggregate p-values using the average of the Fisher aggregation value and the mean aggregation method.
}
