% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_tip.R
\name{calc_ambiguous_above}
\alias{calc_ambiguous_above}
\title{Calculate a line that is parallel to the k-means line and a distance from it that is equal to 66\% of the distance
between the droplets with the biggest library size and the k-means line. The area bound by the k-means line and this
line is referred to as ambiguous area and in it droplets are rejected by thresholding FDR_multi.
Above this ambiguous area, droplets are assumed to contain nuclei/cells and their FDR_multi is set to 0 by default.}
\usage{
calc_ambiguous_above(data, equation)
}
\arguments{
\item{data}{Dataframe of droplets with columns Total_RNA and Total_chromatin.}

\item{equation}{2d vector c(intercept, slope) defining the k-means line.}
}
\value{
2d vector c(intercept, slope) defining the k-means line.
}
\description{
Calculate a line that is parallel to the k-means line and a distance from it that is equal to 66\% of the distance
between the droplets with the biggest library size and the k-means line. The area bound by the k-means line and this
line is referred to as ambiguous area and in it droplets are rejected by thresholding FDR_multi.
Above this ambiguous area, droplets are assumed to contain nuclei/cells and their FDR_multi is set to 0 by default.
}
