% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_tip.R
\name{calc_intercept_of_parallel_line}
\alias{calc_intercept_of_parallel_line}
\title{Calculate the equation of a line that is parallel to a given line and a given distance from it.}
\usage{
calc_intercept_of_parallel_line(equation, dist)
}
\arguments{
\item{equation}{A 2d vector c(intercept, slope) defining a line in 2d Euclidean space.}

\item{dist}{A real number specifying the distance between the given line and the line to be calculated.}
}
\value{
A 2d vector c(intercept, slope) defining a line in 2d Euclidean space
}
\description{
Calculate the equation of a line that is parallel to a given line and a given distance from it.
}
