% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_tip.R
\name{calc_lower_line}
\alias{calc_lower_line}
\title{Calculate a line that is parallel to a given line and passes through the point (x, y).}
\usage{
calc_lower_line(x, y, equation)
}
\arguments{
\item{x}{A real number for the x coordinate of the point.}

\item{y}{A real number for the y coordinate of the point.}

\item{equation}{A 2d vector c(intercept, slope) defining a line in 2d Euclidean space.}
}
\value{
A 2d vector c(intercept, slope) defining a line in 2d Euclidean space.
}
\description{
Calculate a line that is parallel to a given line and passes through the point (x, y).
}
