% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellranger_k_means_v2.R
\name{call_cells}
\alias{call_cells}
\title{Call cells using k-means and store this information in is_cell.}
\usage{
call_cells(data, verbose = TRUE)
}
\arguments{
\item{data}{Dataframe with columns 'atac_count' (int), 'rna_count' (int), 'excluded' (bool), 'is_cell'.}

\item{verbose}{if TRUE various intermediate steps and plots are printed.}
}
\value{
Dataframe with columns 'atac_count' (int), 'rna_count' (int), 'excluded' (bool), 'is_cell'.
}
\description{
Call cells using k-means and store this information in is_cell.
}
