% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eD.R
\name{cell_calling_plot}
\alias{cell_calling_plot}
\title{Creates the cell calling plot.}
\usage{
cell_calling_plot(x_axis, y_axis, colors, title)
}
\arguments{
\item{x_axis}{Vector with values for the x-coordinate.}

\item{y_axis}{Vector with values for the y-coordinate.}

\item{colors}{Vector with values for the color to be used in the scatter plot.}

\item{title}{character to be used as title for the plot.}
}
\value{
invisible(NULL)
}
\description{
Creates the cell calling plot.
}
