% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_tip.R
\name{dist_point_to_line}
\alias{dist_point_to_line}
\title{Calculates the distance between a point and a line.}
\usage{
dist_point_to_line(equation, point)
}
\arguments{
\item{equation}{A vector c(intercept, slope) that defines a line in 2d Euclidean space.}

\item{point}{A 2d vector of coordinates.}
}
\value{
A single real number.
}
\description{
Calculates the distance between a point and a line.
}
