% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emptydrops_with_barhop.R
\name{emptydrops_with_barhop}
\alias{emptydrops_with_barhop}
\title{Calls cells using emptydrops on a single modality. It models the soup using droplets with counts between barhop_end and lower.}
\usage{
emptydrops_with_barhop(count_matrix, lower, barhop_end, niters = 10000)
}
\arguments{
\item{count_matrix}{dgCMatrix with the counts.}

\item{lower}{single real number at or below which droplets are used to model the RNA soup.}

\item{barhop_end}{single real number below which droplets are assumed to be full of barhops and they are excluded from modeling the soup.}

\item{niters}{the number of iterations to use for the simulations that compare the profile of each droplet to the ambient profile.}
}
\value{
a DataFrame (in the format of the emptyDrops outputs) with the following components:
Total*:
Integer, the total count for each barcode.
LogProb:
Numeric, the log-probability of observing the barcode's count vector under the null model.
PValue:
Numeric, the Monte Carlo p-value against the null model.
Limited:
Logical, indicating whether a lower p-value could be obtained by increasing niters.
FDR:
Numeric, the p-values corrected using the Benjamini-Hochberg method
}
\description{
Calls cells using emptydrops on a single modality. It models the soup using droplets with counts between barhop_end and lower.
}
