% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_3_normals.R
\name{equil_of_normals}
\alias{equil_of_normals}
\title{Find the point of equal probability between two Gaussian components of a Gaussian Mixture Model}
\usage{
equil_of_normals(mu1, sd1, lam1, mu2, sd2, lam2)
}
\arguments{
\item{mu1}{the mean of the first Gaussian distribution}

\item{sd1}{the standard deviation of the first Gaussian distribution}

\item{lam1}{the mixing ratio of the first Gaussian distribution}

\item{mu2}{the mean of the second Gaussian distribution}

\item{sd2}{the standard deviation of the second Gaussian distribution}

\item{lam2}{the mixing ratio of the second Gaussian distribution}
}
\description{
Find the point of equal probability between two Gaussian components of a Gaussian Mixture Model
}
