% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_tip.R
\name{find_tip}
\alias{find_tip}
\title{Finds the droplets that are in 1) the top 1\% of RNA counts and in the top 0.1\% of atac counts, and then
calculates their mean RNA counts and mean ATAC counts.}
\usage{
find_tip(data)
}
\arguments{
\item{data}{Dataframe with data$rna being the number of RNA counts, and data$atac being the number of ATAC counts.}
}
\value{
A real vector with the two mean values.
}
\description{
Finds the droplets that are in 1) the top 1\% of RNA counts and in the top 0.1\% of atac counts, and then
calculates their mean RNA counts and mean ATAC counts.
}
