% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher.R
\name{fisher_v2}
\alias{fisher_v2}
\title{Use chi-squared function to aggregate p-values derived from different hypothesis tests}
\usage{
fisher_v2(pvalues)
}
\arguments{
\item{pvalues}{Vector of real numbers.}
}
\value{
Real number of the aggregated p-value.
}
\description{
Use chi-squared function to aggregate p-values derived from different hypothesis tests
}
