% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_3_normals.R
\name{fit_3_normals}
\alias{fit_3_normals}
\title{Fits a mixture of three normal distributions to histogram of RNA counts, to infer the parameters lower and barhop_end
which are used in emptydrops.}
\usage{
fit_3_normals(exp_counts_per_cell)
}
\arguments{
\item{exp_counts_per_cell}{a vector with RNA library size per barcode.}
}
\value{
A 2d vector with the values lower and barhop_end for RNA.
}
\description{
Fits a mixture of three normal distributions to histogram of RNA counts, to infer the parameters lower and barhop_end
which are used in emptydrops.
}
