% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_3_normals.R
\name{fit_3_normals_atac}
\alias{fit_3_normals_atac}
\title{Fits a mixture of three normal distributions to histogram of ATAC counts, to infer the parameters lower and barhop_end
which are used in emptydrops.}
\usage{
fit_3_normals_atac(exp_counts_per_cell, verbose = TRUE)
}
\arguments{
\item{exp_counts_per_cell}{a vector with the ATAC library size per barcode.}

\item{verbose}{if TRUE various intermediate steps and plots are printed.}
}
\value{
A 2d vector with the values lower and barhop_end for ATAC.
}
\description{
Fits a mixture of three normal distributions to histogram of ATAC counts, to infer the parameters lower and barhop_end
which are used in emptydrops.
}
