% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellranger_k_means_v2.R
\name{log_transform}
\alias{log_transform}
\title{Log-transform the data.}
\usage{
log_transform(data)
}
\arguments{
\item{data}{Dataframe with column atac_count and column rna_count}
}
\value{
Dataframe with with log-transformed column values
}
\description{
Log-transform the data.
}
