% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_perp_bisector.R
\name{log_transform_v2}
\alias{log_transform_v2}
\title{log-transform the data}
\usage{
log_transform_v2(data)
}
\arguments{
\item{data}{Dataframe with column x and column y}
}
\value{
Dataframe with with log-transformed column values
}
\description{
log-transform the data
}
