% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher.R
\name{mean_aggr}
\alias{mean_aggr}
\title{Mean aggregation of real values.}
\usage{
mean_aggr(pvalues, r)
}
\arguments{
\item{pvalues}{Vector of real numbers.}

\item{r}{Real number for the kind of mean which will be used to aggregate the p-values.
For instance, r=1 is the arithmetic mean, r=-1 is the harmonic mean etc.}
}
\value{
Real number of the aggregated p-value.
}
\description{
Mean aggregation of real values.
}
