% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_perp_bisector.R
\name{perp_bisector}
\alias{perp_bisector}
\title{Calculate the perpendicular bisector line to the line segment defined by two points in 2d Euclidean space.}
\usage{
perp_bisector(center1, center2)
}
\arguments{
\item{center1}{A point in 2d Euclidean space.}

\item{center2}{A point in 2d Euclidean space.}
}
\value{
Two dimensional vector c(intercept, slope)
}
\description{
Calculate the perpendicular bisector line to the line segment defined by two points in 2d Euclidean space.
}
