% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_3_normals.R
\name{plot_counts}
\alias{plot_counts}
\title{Plot histogram of counts for lower and barhop_end inference}
\usage{
plot_counts(
  nCount,
  mu1,
  sigma1,
  lam1,
  mu2,
  sigma2,
  lam2,
  max = 50000,
  vline1,
  vline2,
  label
)
}
\arguments{
\item{nCount}{vector with the RNA or ATAC counts}

\item{mu1}{mu of first Gaussian}

\item{sigma1}{sigma of first Gaussian}

\item{lam1}{lambda of first Gaussian}

\item{mu2}{mu of first Gaussian}

\item{sigma2}{sigma of first Gaussian}

\item{lam2}{lambda of first Gaussian}

\item{max}{upper limit of RNA range}

\item{vline1}{x value of first vertical line}

\item{vline2}{x value of second vertical line}

\item{label}{label for x axis}
}
\value{
None
}
\description{
Plot histogram of counts for lower and barhop_end inference
}
