% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellranger_k_means_v2.R
\name{simpler_ordmag}
\alias{simpler_ordmag}
\title{Finds a nice initialization for the K-means clustering}
\usage{
simpler_ordmag(x)
}
\arguments{
\item{x}{Vector of real values.}
}
\value{
Value roughly equal to 1/10 of the 99\% percentile of values.
}
\description{
Finds a nice initialization for the K-means clustering
}
