% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellranger_k_means_v2.R
\name{transfer_labels_to_dedupl_v3}
\alias{transfer_labels_to_dedupl_v3}
\title{Assigns the value is_cell of the unique elements to the value is_cell of their corresponding duplicates.}
\usage{
transfer_labels_to_dedupl_v3(data)
}
\arguments{
\item{data}{Dataframe with columns 'atac_count' (int), 'rna_count' (int), 'excluded' (bool), 'is_cell' (bool), 'dup' (bool).
data is sorted wrt 'atac_count' (int), 'rna_count' (int), 'dup' (bool).}
}
\value{
Dataframe with altered is_cell column.
}
\description{
Assigns the value is_cell of the unique elements to the value is_cell of their corresponding duplicates.
}
